requirements.txt
ipykernel
langchain
langchainhub
langchain-community
langchain-huggingface
langgraph
langgraph-checkpoint-sqlite
python-dotenv
text-generation
transformers